Bug fix. Patch Metrics would sometimes add the MDCP field when the calculate option was not selected. Changed the field types for Patch Metrics so that larger result numbers could be accommodated particularly for number of patches. This became an issue when 1-m resolution land cover was being used and small patch sizes were defined. Cleaned up log.py and raster.py by removing old unused code.

+++ ToolboxSource/ATtILA2/ATtILA2/utils/calculate.py
@@ -816,8 +816,8 @@ def getMDCP(outIdField, newTable, mdcpDict, optionalGroupsList, outClassName):
             del outTableRows, outTableRow
         # If QA fields are selected, add fields for pwn (patches w/ neighbors) and pwon (patches w/o) for the class
         if globalConstants.qaCheckName in optionalGroupsList:
-            arcpy.AddField_management(newTable, outClassName+"_PWN", "Double")
-            arcpy.AddField_management(newTable, outClassName+"_PWON", "Double")
+            arcpy.AddField_management(newTable, outClassName+"_PWN", "LONG")
+            arcpy.AddField_management(newTable, outClassName+"_PWON", "LONG")
 
         mdcpFieldName = outClassName+"_MDCP"
